requirements.txt
pandas>=1.5
pyarrow>=14.0
matplotlib>=3.6
requests>=2.28
tenacity>=8.0
python-dotenv>=0.21
pyyaml>=6.0
pytest>=7.0
beautifulsoup4>=4.12
lxml>=4.9
contourpy==1.3.3
cycler==0.12.1
fonttools==4.61.1
kiwisolver==1.4.9
matplotlib==3.10.8
numpy==2.4.2
packaging==26.0
pandas==3.0.0
pillow==12.1.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
six==1.17.0
